Bug #69979. [Core] Type overflow in a SqliteTypesConverter

+++ clearth-core/src/main/java/com/exactprosystems/clearth/utils/tabledata/typing/converter/SqliteTypesConverter.java
@@ -33,7 +33,7 @@ public class SqliteTypesConverter implements DbTypesConverter
 			case BIGINT:
 			case TINYINT:
 			case SMALLINT:
-				return TableDataType.INTEGER;
+				return TableDataType.LONG;
 			case DATE:
 				return TableDataType.LOCALDATETIME;
 			case NUMERIC:
